@@ -1,16 +1,15 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+// import React from 'react'
+// import ReactDOM from 'react-dom'
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 import App from './App'
 import * as serviceWorker from './serviceWorker'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const container = document.getElementById('root') as HTMLDivElement
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<StrictMode><App /></StrictMode>)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+serviceWorker.unregister()
