@@ -1,5 +1,5 @@
 import React from 'react'
-import { BrowserRouter as Router } from 'react-router-dom'
+import { BrowserRouter } from 'react-router-dom'
 import { CssBaseline } from '@material-ui/core'
 import { useInit } from '../../lib/init'
 
@@ -10,8 +10,8 @@ export default function Main() {
   useInit()
   return <>
     <CssBaseline />
-    <Router>
+    <BrowserRouter>
       <Routes />
-    </Router>
+    </BrowserRouter>
   </>
 }
